Add asset re-use support to devtest.sh.
We can now re-use the images/code from a previous run to significantly
reduce the runtime of a devtest.sh run.

+++ scripts/devtest_overcloud.sh
@@ -27,11 +27,13 @@ USERS=${9:-${USERS:-''}}
 ##    used to pass additional build-time specific arguments to disk-image-create.
 ##    ::
 
-$TRIPLEO_ROOT/diskimage-builder/bin/disk-image-create $NODE_DIST \
-    -a $NODE_ARCH -o $TRIPLEO_ROOT/overcloud-control \
-    boot-stack cinder os-collect-config neutron-network-node \
-    dhcp-all-interfaces stackuser swift-proxy swift-storage ${OVERCLOUD_DIB_EXTRA_ARGS:-} 2>&1 | \
-    tee $TRIPLEO_ROOT/dib-overcloud-control.log
+if [ ! -e $TRIPLEO_ROOT/overcloud-control.qcow2 -o "$USE_CACHE" == "0" ] ; then #nodocs
+    $TRIPLEO_ROOT/diskimage-builder/bin/disk-image-create $NODE_DIST \
+        -a $NODE_ARCH -o $TRIPLEO_ROOT/overcloud-control \
+        boot-stack cinder os-collect-config neutron-network-node \
+        dhcp-all-interfaces stackuser swift-proxy swift-storage ${OVERCLOUD_DIB_EXTRA_ARGS:-} 2>&1 | \
+        tee $TRIPLEO_ROOT/dib-overcloud-control.log
+fi #nodocs
 
 ## #. Load the image into Glance:
 ##    ::
@@ -44,11 +46,13 @@ load-image -d $TRIPLEO_ROOT/overcloud-control.qcow2
 ##    network.
 ##    ::
 
-$TRIPLEO_ROOT/diskimage-builder/bin/disk-image-create $NODE_DIST \
-    -a $NODE_ARCH -o $TRIPLEO_ROOT/overcloud-compute \
-    nova-compute nova-kvm neutron-openvswitch-agent os-collect-config \
-    dhcp-all-interfaces stackuser ${OVERCLOUD_DIB_EXTRA_ARGS:-} 2>&1 | \
-    tee $TRIPLEO_ROOT/dib-overcloud-compute.log
+if [ ! -e $TRIPLEO_ROOT/overcloud-compute.qcow2 -o "$USE_CACHE" == "0" ] ; then #nodocs
+    $TRIPLEO_ROOT/diskimage-builder/bin/disk-image-create $NODE_DIST \
+        -a $NODE_ARCH -o $TRIPLEO_ROOT/overcloud-compute \
+        nova-compute nova-kvm neutron-openvswitch-agent os-collect-config \
+        dhcp-all-interfaces stackuser ${OVERCLOUD_DIB_EXTRA_ARGS:-} 2>&1 | \
+        tee $TRIPLEO_ROOT/dib-overcloud-compute.log
+fi #nodocs
 
 ## #. Load the image into Glance:
 ##    ::
@@ -111,8 +115,10 @@ heat stack-create -f $TRIPLEO_ROOT/tripleo-heat-templates/overcloud.yaml \
 ## #. While we wait for the stack to come up, build an end user disk image and
 ##    register it with glance.::
 
-$TRIPLEO_ROOT/diskimage-builder/bin/disk-image-create $NODE_DIST vm \
-    -a $NODE_ARCH -o $TRIPLEO_ROOT/user 2>&1 | tee $TRIPLEO_ROOT/dib-user.log
+if [ ! -e $TRIPLEO_ROOT/user.qcow2 -o "$USE_CACHE" == "0" ] ; then #nodocs
+    $TRIPLEO_ROOT/diskimage-builder/bin/disk-image-create $NODE_DIST vm \
+        -a $NODE_ARCH -o $TRIPLEO_ROOT/user 2>&1 | tee $TRIPLEO_ROOT/dib-user.log
+fi #nodocs
 
 ## #. Get the overcloud IP from 'nova list'
 ##    ::
